타입정리 / state 정리 / styled 정리 / msw handler추가
