+++ actions/shop.ts
@@ -1,0 +1,16 @@
+import { createAsyncThunk } from "@reduxjs/toolkit";
+import api from "../util/api";
+
+const BASE_URL = process.env.NEXT_PUBLIC_API_URL;
+
+export const getCategoryList = createAsyncThunk(
+  "shop/getCategoryList",
+  async (data, { rejectWithValue }) => {
+    try {
+      const response = await api.get(BASE_URL + "/api/v1/shop/category");
+      return response.data;
+    } catch (error: any) {
+      return rejectWithValue(error.response.data);
+    }
+  }
+);
