+++ mocks/mockData/product.ts
@@ -1,0 +1,273 @@
+import Product1 from "../../assets/img/temp/product1.png";
+import ProductList1 from "../../assets/img/temp/productList1.png";
+import ProductDetail1 from "../../assets/img/temp/product_detail.png";
+import ProductDetail2 from "../../assets/img/temp/product_detail2.png";
+import ProductDetail3 from "../../assets/img/temp/product_detail3.png";
+import ProductDetail4 from "../../assets/img/temp/product_detail4.png";
+import ProductDetail5 from "../../assets/img/temp/product_detail5.png";
+import ProductDetail6 from "../../assets/img/temp/product_detail6.png";
+import ProductDetail7 from "../../assets/img/temp/product_detail7.png";
+import ProductDesc from "../../assets/img/temp/product_desc.png";
+
+const product1 = {
+  id: 1,
+  img: [ProductList1.src, Product1.src],
+  label: "무료배송",
+  like: false,
+  soldOut: false,
+  rank: 1,
+  brand: "HAEUL",
+  name: "터널 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 10,
+  price: 11000,
+  categoryCode: 4,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 22000,
+};
+const product2 = {
+  id: 2,
+  img: [ProductList1.src, Product1.src],
+  label: "당일배송",
+  like: false,
+  soldOut: true,
+  rank: 2,
+  brand: "HAEUL",
+  name: "터널 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 40,
+  price: 33000,
+  categoryCode: 4,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 66000,
+};
+const product3 = {
+  id: 3,
+  img: [ProductList1.src, Product1.src],
+  label: "",
+  like: false,
+  soldOut: false,
+  rank: 3,
+  brand: "HAEUL",
+  name: "하울 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 15,
+  price: 22900,
+  categoryCode: 6,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 33000,
+};
+const product4 = {
+  id: 4,
+  img: [ProductList1.src, Product1.src],
+  label: "",
+  like: false,
+  soldOut: false,
+  rank: 4,
+  brand: "HAEUL",
+  name: "원터치 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 20,
+  price: 77000,
+  categoryCode: 1,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 99000,
+};
+const product5 = {
+  id: 5,
+  img: [ProductList1.src, Product1.src],
+  label: "마감임박",
+  like: false,
+  soldOut: false,
+  rank: 5,
+  brand: "HAEUL",
+  name: "원터치 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 80,
+  price: 22000,
+  categoryCode: 1,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 55000,
+};
+const product6 = {
+  id: 6,
+  img: [ProductList1.src, Product1.src],
+  label: "",
+  like: false,
+  soldOut: false,
+  rank: 6,
+  brand: "HAEUL",
+  name: "돔 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 55,
+  price: 9900,
+  categoryCode: 2,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 20000,
+};
+const product7 = {
+  id: 7,
+  img: [ProductList1.src, Product1.src],
+  label: "",
+  like: false,
+  soldOut: true,
+  rank: 7,
+  brand: "HAEUL",
+  name: "돔 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 10,
+  price: 32000,
+  categoryCode: 2,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 52000,
+};
+const product8 = {
+  id: 8,
+  img: [ProductList1.src, Product1.src],
+  label: "무료배송",
+  like: false,
+  soldOut: false,
+  rank: 8,
+  brand: "HAEUL",
+  name: "리빙쉘 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 20,
+  price: 44400,
+  categoryCode: 3,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 62000,
+};
+const product9 = {
+  id: 9,
+  img: [ProductList1.src, Product1.src],
+  label: "",
+  like: false,
+  soldOut: false,
+  rank: 9,
+  brand: "HAEUL",
+  name: "리빙쉘 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 30,
+  price: 41000,
+  categoryCode: 3,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 62000,
+};
+const product10 = {
+  id: 10,
+  img: [ProductList1.src, Product1.src],
+  label: "무료배송",
+  like: false,
+  soldOut: false,
+  rank: 10,
+  brand: "HAEUL",
+  name: "하울 너도밤나무 파인우드행어 원목 감성캠핑용품 인디언행어",
+  dcRate: 22,
+  price: 9500,
+  categoryCode: 6,
+  detailImg: [
+    ProductDetail1.src,
+    ProductDetail2.src,
+    ProductDetail3.src,
+    ProductDetail4.src,
+    ProductDetail5.src,
+    ProductDetail6.src,
+    ProductDetail7.src,
+  ],
+  desc: ProductDesc.src,
+  originPrice: 32000,
+};
+
+export const mockProductList = [
+  product1,
+  product2,
+  product3,
+  product4,
+  product5,
+  product6,
+  product7,
+  product8,
+  product9,
+  product10,
+];
+
+export const mockCategoryList = [
+  { id: 1, name: "원터치 텐트" },
+  { id: 2, name: "돔 텐트" },
+  { id: 3, name: "리빙쉘 텐트" },
+  { id: 4, name: "터널 텐트" },
+  { id: 5, name: "차박 텐트" },
+  { id: 6, name: "하울 텐트" },
+];
