+++ mocks/mockData/shop.ts
@@ -1,0 +1,23 @@
+import Tent from "../../assets/img/product/category/tent.png";
+import Tarp from "../../assets/img/product/category/tarp.png";
+import SleepingBag from "../../assets/img/product/category/sleepingBag.png";
+import Table from "../../assets/img/product/category/table.png";
+import Chair from "../../assets/img/product/category/chair.png";
+import Mat from "../../assets/img/product/category/mat.png";
+import Coppell from "../../assets/img/product/category/coppell.png";
+import Fire from "../../assets/img/product/category/fire.png";
+import Lantern from "../../assets/img/product/category/lantern.png";
+import BoxImg from "../../assets/img/product/category/box.png";
+
+export const mockShopCategoryList = [
+  { id: 0, name: "텐트", img: Tent.src },
+  { id: 1, name: "타프", img: Tarp.src },
+  { id: 2, name: "침낭", img: SleepingBag.src },
+  { id: 3, name: "테이블", img: Table.src },
+  { id: 4, name: "의자", img: Chair.src },
+  { id: 5, name: "매트", img: Mat.src },
+  { id: 6, name: "코펠", img: Coppell.src },
+  { id: 7, name: "화로", img: Fire.src },
+  { id: 8, name: "랜턴", img: Lantern.src },
+  { id: 9, name: "수납", img: BoxImg.src },
+];
