+++ pages/shop/product/index.tsx
@@ -1,8 +1,6 @@
-import React, { FC, useEffect, useState } from "react";
+import React, { FC, useEffect } from "react";
 import ProductLayout from "@layout/ProductLayout";
 import Grid from "@mui/material/Grid";
-
-import styled from "@emotion/styled";
 import ProductLike from "../../../assets/img/temp/productLike.svg";
 import ProductDislike from "../../../assets/img/temp/productDislike.svg";
 import { NumberCommaFilter } from "../../../util/commonFilter";
@@ -15,215 +13,67 @@ import MenuItem from "@mui/material/MenuItem";
 import { Select } from "@mui/material";
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import { SelectChangeEvent } from "@mui/material/Select";
-import { getList } from "../../../actions/product";
+import { getCategoryList, getList } from "../../../actions/product";
 import { useDispatch, useSelector } from "react-redux";
-import { ReqDto } from "../../../type/common/common";
+import {
+  Category,
+  Product,
+  ProductReqData,
+} from "../../../type/product/product";
 import productSlice from "@reducers/product";
+import { AppDispatch } from "../../../store/configureStore";
 
-const ProductImageStyle = styled.div`
-  text-align: center;
-  position: relative;
-  .image {
-    border-radius: 18px;
-  }
-  .label {
-    position: absolute;
-    top: 6%;
-    left: 11%;
-    background-color: #5bbd6b;
-    color: #ffffff;
-    border-radius: 4px;
-    font-size: 12px;
-    height: 22px;
-    width: 54px;
-    line-height: 20px;
-    text-align: center;
-    font-weight: 800;
-  }
-  .like {
-    cursor: pointer;
-    position: absolute;
-    top: 82%;
-    left: 79%;
-    z-index: 1;
-  }
-  .sold-out {
-    position: absolute;
-    border-radius: 18px;
-    top: 0%;
-    left: 5%;
-    width: 289px;
-    height: 289px;
-    background-color: rgba(0, 0, 0, 0.5);
-    .text {
-      position: absolute;
-      top: 44%;
-      left: 42%;
-      color: #ffffff;
-      font-size: 20px;
-      font-weight: 700;
-    }
-  }
-`;
+//styled-component
+import {
+  ProductImageStyle,
+  ProductInfoGrid,
+  ProductListInfoGrid,
+} from "../../../assets/styles/styled/product/productList";
 
-const ProductListInfoGrid = styled(Grid)`
-  .grid-product-text-padding {
-    padding-left: 24px;
-    p {
-      font-weight: 400;
-      font-size: 16px;
-    }
-  }
-  .grid-product-sort-margin {
-    margin-bottom: 24px;
-    text-align: center;
-    .MuiSelect-select.MuiSelect-outlined.MuiInputBase-input.MuiOutlinedInput-input {
-      padding: 0px 50px 0px 0px;
+const ProductComponent: FC = () => {
+  const dispatch = useDispatch<AppDispatch>();
 
-      font-weight: 400;
-      font-size: 16px;
-    }
-    .MuiOutlinedInput-notchedOutline {
-      border: none;
-    }
-  }
-
-  .grid-product-filter-margin {
-    padding-left: 15px;
-    text-align: center;
-    margin-bottom: 40px;
-    .MuiOutlinedInput-root.MuiOutlinedInput-root.MuiInputBase-colorPrimary {
-      border-radius: 25px !important;
-      .MuiSelect-select.MuiSelect-outlined.MuiInputBase-input.MuiOutlinedInput-input {
-        padding: 10px 30px 10px 20px;
-        font-weight: 400;
-        font-size: 16px;
-        color: #919191;
-      }
-    }
-  }
-`;
-
-const ProductInfoGrid = styled(Grid)`
-  .grid-product-img-margin {
-    margin-bottom: 12px;
-  }
-  .div-product-desc {
-    margin-left: 17px;
-    margin-bottom: 30px;
-    div {
-      max-width: 95% !important;
-    }
-    .grid-product-brand {
-      margin-bottom: 12px;
-      .rank {
-        font-size: 12pt;
-        color: #424242;
-        margin-right: 7px;
-      }
-      .brand {
-        font-size: 12pt;
-        color: #919191;
-      }
-    }
-    .grid-product-name {
-      margin-bottom: 12px;
-      p {
-        color: #383838;
-        fontsize: 16px;
-      }
-    }
-    .grid-product-price {
-      display: flex;
-      margin-bottom: 12px;
-      .dcRate {
-        font-size: 20pt;
-        color: #fc6e51;
-        margin-right: 15px;
-      }
-      .price {
-        font-size: 20pt;
-        color: #222222;
-      }
-    }
-  }
-`;
-
-type Product = {
-  id: number;
-  img: string[];
-  label: string;
-  like: boolean;
-  soldOut: boolean;
-  rank: number | null;
-  brand: string;
-  name: string;
-  dcRate: number;
-  price: number;
-  category: string;
-};
-const Product: FC = () => {
-  const dispatch = useDispatch();
   const router = useRouter();
 
   const query = router.query;
 
-  const [selectedCategory, setSelectedCategory] = useState("");
-
-  //temp categoryList
-  const categoryList = [
-    "원터치 텐트",
-    "돔 텐트",
-    "리빙쉘 텐트",
-    "터널 텐트",
-    "차박 텐트",
-    "하울",
-    "원터치 텐트",
-    "돔 텐트",
-    "리빙쉘 텐트",
-    "터널 텐트",
-    "차박 텐트",
-    "하울",
-  ];
-
-  const stateProductList: Product[] = useSelector(
+  const productList: Product[] = useSelector(
     (state: any) => state.product.productList
   );
 
-  const getListDone: Boolean = useSelector(
-    (state: any) => state.product.getListDone
+  const categoryList: Category[] = useSelector(
+    (state: any) => state.product.categoryList
   );
 
-  const productReqData: any = {
+  const selectedSort: string = useSelector(
+    (state: any) => state.product.selectedSort
+  );
+
+  const productReqData: ProductReqData = {
     page: 0,
     size: 10,
     keyword: "",
     startDate: "",
     endDate: "",
     isList: false,
+    sort: "",
+    category: 0,
   };
 
-  const [productList, setProductList] = useState<Product[]>([]);
-  const [selectedSort, setSelectedSort] = React.useState("");
+  useEffect(() => {
+    productReqData.isList = true;
+    // @ts-ignore
+    dispatch(getList(productReqData));
+    dispatch(getCategoryList());
+  }, []);
 
   const handleChangeSort = (e: SelectChangeEvent) => {
-    setSelectedSort(e.target.value);
+    dispatch(productSlice.actions.setSelectedSort(e.target.value));
     productReqData.isList = true;
     productReqData.sort = e.target.value;
 
-    // @ts-ignore
     dispatch(getList(productReqData));
   };
-  useEffect(() => {
-    setSelectedCategory(categoryList[0]);
-    productReqData.isList = true;
-    // @ts-ignore
-    dispatch(getList(productReqData));
-  }, []);
-
-  useEffect(() => {
-    setProductList(stateProductList);
-  }, [getListDone]);
 
   const handleClickLike = (selectedId: number) => {
     const mappedProductList = productList.map((m: Product) => {
@@ -232,7 +82,7 @@ const Product: FC = () => {
         like: selectedId === m.id ? !m.like : m.like,
       };
     });
-    setProductList(mappedProductList);
+    dispatch(productSlice.actions.setProductList(mappedProductList));
   };
 
   const [filter1, setFilter1] = React.useState("");
@@ -253,7 +103,7 @@ const Product: FC = () => {
     <ProductLayout category={query.category}>
       <CategoryTab
         categoryList={categoryList}
-        setSelectedCategory={setSelectedCategory}
+        productReqData={productReqData}
       />
       <ProductListInfoGrid container>
         <Grid item container xs={12} className={"grid-product-filter-margin"}>
@@ -409,4 +259,4 @@ const Product: FC = () => {
   );
 };
 
-export default Product;
+export default ProductComponent;
