+++ reducers/index.ts
@@ -5,6 +5,7 @@ import authSlice from "./auth";
 import accountSlice from "./account";
 import productSlice from "./product";
 import communitySlice from "./community";
+import shopSlice from "./shop";
 
 // (이전상태, 액션) => 다음상태
 const rootReducer = (state: any, action: any) => {
@@ -17,6 +18,7 @@ const rootReducer = (state: any, action: any) => {
         account: accountSlice.reducer,
         product: productSlice.reducer,
         community: communitySlice.reducer,
+        shop: shopSlice.reducer,
       });
       return combinedReducer(state, action);
     }
