+++ type/shop/shop.ts
@@ -1,0 +1,4 @@
+export type Category = {
+  id: number;
+  name: string;
+};
